+++ docs/CHANGELOG.md
@@ -4,6 +4,19 @@
 
 ## [Unreleased]
 
+## [2.2.1] - 2020-09-01
+### Optimized
+- code organization
+### Added
+- games
+- diagnose script in ext/init
+- emoji command to peek emojis from other servers
+- tutorial command
+### Removed
+- since removing issue command affect specific features, issue command will be left
+### Fixed
+- help command now identifies optional params correctly
+
 ## [2.2.0] - 2020-08-16
 ### Added
 - some commands now contains multi-line help messages
@@ -236,7 +249,8 @@
 - no more verbose mode :O
 
 
-[Unreleased]:   https://github.com/windowsboy111/Merlin-py/compare/2.2.0...HEAD
+[Unreleased]:   https://github.com/windowsboy111/Merlin-py/compare/2.2.1...HEAD
+[2.2.1]:        https://github.com/windowsboy111/Merlin-py/compare/2.2.0...2.2.1
 [2.2.0]:        https://github.com/windowsboy111/Merlin-py/compare/2.2-b1...2.2.0
 [2.2-b1]:       https://github.com/windowsboy111/Merlin-py/compare/2.2-b...2.2-b1
 [2.2-b]:        https://github.com/windowsboy111/Merlin-py/compare/2.2.a...2.2-b